% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moderated_ttest.R
\name{mttest}
\alias{mttest}
\alias{calculate_moderated_ttest}
\alias{calculate_moderated_ttest_hgnc}
\alias{calculate_moderated_ttest_uniprot}
\title{Moderated t-test for proteomics}
\usage{
calculate_moderated_ttest(input_file)

calculate_moderated_ttest_hgnc(input_file)

calculate_moderated_ttest_uniprot(input_file)
}
\arguments{
\item{input_file}{a data.frame containing at least three columns, 
gene, replicate 1 and replicate 2. Triplicates and so on can be 
also  be included. The function assumes that the first column 
contains gene information.}
}
\description{
Calculate the moderated ttest using limma package.
}
\seealso{
Other statistics: 
\code{\link{normalize}()},
\code{\link{scatter}}
}
\author{
April/Frederik
}
\concept{statistics}
