% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_files.R
\name{compare_two_files}
\alias{compare_two_files}
\alias{compare_two_files_a}
\alias{compare_two_files_a_scatter}
\alias{compare_two_files_b}
\alias{compare_two_files_b_scatter}
\alias{compare_two_files_aa}
\alias{compare_two_files_aa_scatter}
\alias{compare_two_files_bb}
\alias{compare_two_files_bb_scatter}
\alias{compare_two_files_cc}
\alias{compare_two_files_cc_scatter}
\alias{compare_two_files_pf_a}
\alias{compare_two_files_pf_a_size}
\alias{compare_two_files_pf_b}
\alias{compare_two_files_pf_b_size}
\alias{compare_two_files_pf_aa}
\alias{compare_two_files_pf_aa_size}
\alias{compare_two_files_pf_bb}
\alias{compare_two_files_pf_bb_size}
\alias{compare_two_files_pf_cc}
\alias{compare_two_files_pf_cc_size}
\title{Compare files}
\usage{
compare_two_files_a(orig, subset, overlaps)

compare_two_files_a_scatter(orig, subset, overlaps)

compare_two_files_b(orig, subset, overlaps)

compare_two_files_b_scatter(orig, subset, overlaps)

compare_two_files_aa(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_aa_scatter(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_bb(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_bb_scatter(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_cc(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_cc_scatter(orig, subset, overlaps1, overlaps2, overlaps3)

compare_two_files_pf_a(orig, subset, overlaps)

compare_two_files_pf_a_size(orig, subset, overlaps, increase)

compare_two_files_pf_b(orig, subset, overlaps)

compare_two_files_pf_b_size(orig, subset, overlaps, increase)

compare_two_files_pf_aa(orig, subset, o_12, o_13, o_123)

compare_two_files_pf_aa_size(orig, subset, o_12, o_13, o_123, increase)

compare_two_files_pf_bb(orig, subset, o_21, o_213)

compare_two_files_pf_bb_size(orig, subset, o_21, o_23, o_213, increase)

compare_two_files_pf_cc(orig, subset, o_31, o_32, o_312)

compare_two_files_pf_cc_size(orig, subset, o_31, o_32, o_312, increase)
}
\description{
Compare files
}
\author{
April/Frederik
}
\concept{infile}
