% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotFamilies_1quadrant.R
\name{makePlotFamilies_1quadrant}
\alias{makePlotFamilies_1quadrant}
\title{Make plot families in first quadrant.}
\usage{
makePlotFamilies_1quadrant(data_fam, data, sortPF)
}
\description{
Make plot families in first quadrant.
}
\author{
April/Frederik
}
\concept{plot}
