% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(table = NULL, type = "median")
}
\arguments{
\item{table}{a data.frame/table.}

\item{type}{what type of normalization? Can be any function
that processes numeric vectors. Default is median.}
}
\description{
Normalize a table of label or label free data. 
The function will calculate the \code{type} e.g. median of
all numeric columns and subtract the median from the 
intensity value.
}
\seealso{
Other statistics: 
\code{\link{mttest}},
\code{\link{scatter}}
}
\author{
Frederik
}
\concept{statistics}
