% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{scatter}
\alias{scatter}
\alias{plot_scatter}
\alias{plot_scatter_qc_gg}
\alias{plot_scatter_exac_gg}
\alias{plot_scatter_qc}
\alias{plot_scatter_multiple_cond}
\alias{plot_scatter_white}
\alias{plot_scatter_exac}
\alias{plot_scatter_exac_multi}
\title{Scatter plots for proteomics data}
\usage{
plot_scatter(d)

plot_scatter_qc_gg(d)

plot_scatter_exac_gg(b, a, n)

plot_scatter_qc(orig, d)

plot_scatter_multiple_cond(orig, d)

plot_scatter_white(orig)

plot_scatter_exac(orig, b, a, n)

plot_scatter_exac_multi(orig, b, a, n)
}
\description{
Make scatter
}
\seealso{
Other statistics: 
\code{\link{mttest}},
\code{\link{normalize}()}
}
\author{
April/Frederik
}
\concept{statistics}
