% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{volcano}
\alias{volcano}
\alias{plot_volcano}
\alias{plot_volcano_qc_gg}
\alias{plot_volcano_qc_gg_1}
\alias{plot_volcano_exac_gg}
\alias{plot_volcano_qc}
\alias{plot_volcano_user_color}
\alias{plot_volcano_multiple_cond}
\alias{plot_volcano_multiple_cond_for_goi}
\alias{plot_volcano_exac}
\alias{plot_volcano_exac_multi}
\title{Volcano plots for proteomics}
\usage{
plot_volcano(d, bait = NULL, main = NULL, sub = NULL)

plot_volcano_qc_gg(d)

plot_volcano_qc_gg_1(data, data1, col, col1, logfc1, logfc2, pval)

plot_volcano_exac_gg(b, a, n)

plot_volcano_qc(d)

plot_volcano_user_color(n_exist, y_exist)

plot_volcano_multiple_cond(d)

plot_volcano_multiple_cond_for_goi(d)

plot_volcano_exac(b, a, n)

plot_volcano_exac_multi(b, a, n)
}
\description{
Will generate a classic volcano plot. \code{plot_volcano} is 
intented for use in R, whereas the remaining functons are used in the shiny
application or intented for the user that requires additional control of 
plotting.
}
\examples{
\dontrun{
# artificial data to illustrate how to use.
data <- data.frame(gene=LETTERS, rep1 = runif(1:26), rep2=rnorm(1:26))
result = calculate_moderated_ttest(data)
plot_volcano(result, bait = 'A', main = 'randomly generated data points')
}
}
\author{
April/Frederik
}
\concept{plotting}
