% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scatter.R
\name{search_scatter}
\alias{search_scatter}
\alias{search_scatter_gg}
\title{Search scatter plot}
\usage{
search_scatter(p, found)

search_scatter_gg(p, found)
}
\description{
Search scatter plot
}
\seealso{
Other search: 
\code{\link{search_volcano}()}
}
\author{
April/Frederik
}
\concept{search}
