% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_volcano.R
\name{search_volcano}
\alias{search_volcano}
\alias{search_volcano_gg}
\alias{search_volcano_gg_reg_label}
\title{Search Volcano plots}
\usage{
search_volcano(p, found)

search_volcano_gg(p, found)

search_volcano_gg_reg_label(p, found)
}
\description{
Search Volcano plots
}
\seealso{
Other search: 
\code{\link{search_scatter}()}
}
\author{
April/Frederik
}
\concept{search}
