% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate_to_groups_for.R
\name{separate_to_groups_for_color}
\alias{separate_to_groups_for_color}
\alias{separate_to_groups_for_color_integrated}
\alias{separate_to_groups_for_cbf_integrated}
\alias{separate_to_groups_for_exac_bar}
\title{Seperate to groups for color}
\usage{
separate_to_groups_for_color(vp_data, color_data, color_theme, type)

separate_to_groups_for_color_integrated(vp_data, threshold, col1, col2)

separate_to_groups_for_cbf_integrated(vp_data, threshold)

separate_to_groups_for_exac_bar(vp_data)
}
\description{
Seperate to groups for color
}
\author{
April/Frederik
}
\concept{grouping}
