% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_layer_for_inweb.R
\name{sp_layer_for_inweb}
\alias{sp_layer_for_inweb}
\alias{sp_layer_for_inweb_no_text}
\alias{sp_layer_for_inweb_cbf}
\alias{sp_layer_for_inweb_cbf_no_text}
\title{InWeb Layers for ggplot2}
\usage{
sp_layer_for_inweb(p, d_in)

sp_layer_for_inweb_no_text(p, d_in)

sp_layer_for_inweb_cbf(p, d_in)

sp_layer_for_inweb_cbf_no_text(p, d_in)
}
\arguments{
\item{p}{plot object}

\item{d_in}{data}
}
\description{
InWeb Layers for ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_snp_to_gene_sgl}()},
\code{\link{sp_layer_for_uploaded_genes}()},
\code{\link{vp_layer_for_inweb}()},
\code{\link{vp_layer_for_snp_to_gene_sgl}()},
\code{\link{vp_layer_for_uploaded_genes_gg}()}
}
\author{
April
}
\concept{layers}
