% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_layer_for_snp_to_gene.R
\name{sp_layer_for_snp_to_gene_sgl}
\alias{sp_layer_for_snp_to_gene_sgl}
\alias{sp_layer_for_snp_to_gene_sgl_no_text}
\alias{sp_layer_for_snp_to_gene_sgl_cbf}
\alias{sp_layer_for_snp_to_gene_sgl_cbf_no_text}
\alias{sp_layer_for_snp_to_gene_mgl}
\alias{sp_layer_for_snp_to_gene_mgl_no_text}
\alias{sp_layer_for_snp_to_gene_mgl_cbf}
\alias{sp_layer_for_snp_to_gene_mgl_cbf_no_text}
\alias{sp_layer_for_snp_to_gene_none}
\alias{sp_layer_for_snp_to_gene_none_cbf}
\title{Layer for mapping uploaded SNPS to genes in ggplot2}
\usage{
sp_layer_for_snp_to_gene_sgl(p, snp_sgl)

sp_layer_for_snp_to_gene_sgl_no_text(p, snp_sgl)

sp_layer_for_snp_to_gene_sgl_cbf(p, snp_sgl)

sp_layer_for_snp_to_gene_sgl_cbf_no_text(p, snp_sgl)

sp_layer_for_snp_to_gene_mgl(p, snp_mgl)

sp_layer_for_snp_to_gene_mgl_no_text(p, snp_mgl)

sp_layer_for_snp_to_gene_mgl_cbf(p, snp_mgl)

sp_layer_for_snp_to_gene_mgl_cbf_no_text(p, snp_mgl)

sp_layer_for_snp_to_gene_none(p, d)

sp_layer_for_snp_to_gene_none_cbf(p, d)
}
\description{
Layer for mapping uploaded SNPS to genes in ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_inweb}()},
\code{\link{sp_layer_for_uploaded_genes}()},
\code{\link{vp_layer_for_inweb}()},
\code{\link{vp_layer_for_snp_to_gene_sgl}()},
\code{\link{vp_layer_for_uploaded_genes_gg}()}
}
\author{
April/Frederik
}
\concept{layers}
