% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_layer_for_uploaded_genes.R
\name{sp_layer_for_uploaded_genes}
\alias{sp_layer_for_uploaded_genes}
\alias{sp_layer_for_uploaded_genes_gg}
\alias{sp_layer_for_uploaded_genes_cbf_gg}
\alias{sp_layer_for_uploaded_genes_no_text}
\alias{sp_layer_for_uploaded_genes_cbf}
\alias{sp_layer_for_uploaded_genes_cbf_no_text}
\alias{sp_layer_for_uploaded_genes_none}
\alias{sp_layer_for_uploaded_genes_none_cbf}
\title{Layer for uploaded genes to ggplot2}
\usage{
sp_layer_for_uploaded_genes(p, d_g2s)

sp_layer_for_uploaded_genes_gg(p, d_g2s, usr_palette, lab)

sp_layer_for_uploaded_genes_cbf_gg(p, d_g2s, lab)

sp_layer_for_uploaded_genes_no_text(p, d_g2s)

sp_layer_for_uploaded_genes_cbf(p, d_g2s)

sp_layer_for_uploaded_genes_cbf_no_text(p, d_g2s)

sp_layer_for_uploaded_genes_none(p, d)

sp_layer_for_uploaded_genes_none_cbf(p, d)
}
\description{
Layer for uploaded genes to ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_inweb}()},
\code{\link{sp_layer_for_snp_to_gene_sgl}()},
\code{\link{vp_layer_for_inweb}()},
\code{\link{vp_layer_for_snp_to_gene_sgl}()},
\code{\link{vp_layer_for_uploaded_genes_gg}()}
}
\author{
April/Frederik
}
\concept{layers}
