% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppi_theme.R, R/theme_genoppi.R
\name{theme_genoppi}
\alias{theme_genoppi}
\title{Genoppi GG theme}
\usage{
theme_genoppi()

theme_genoppi()
}
\description{
Set's up the classic genoppi theme for ggplot2.

Set's up the classic genoppi theme for ggplot2.
}
\author{
April/Frederik

April/Frederik
}
