% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_layer_for_inweb.R
\name{vp_layer_for_inweb}
\alias{vp_layer_for_inweb}
\alias{vp_layer_for_inweb_sf_gg}
\alias{vp_layer_for_inweb_cbf}
\alias{vp_layer_for_inweb_sf}
\alias{vp_layer_for_inweb_no_text_sf}
\alias{vp_layer_for_inweb_cbf_sf}
\alias{vp_layer_for_inweb_cbf_sf_gg}
\alias{vp_layer_for_inweb_cbf_no_text_sf}
\title{InWeb Layers for ggplot2}
\usage{
vp_layer_for_inweb(p, d_in, marker_col, lab)

vp_layer_for_inweb_sf_gg(p, d_in, marker_col, lab)

vp_layer_for_inweb_cbf(p, d_in, lab)

vp_layer_for_inweb_sf(p, d_in, marker_col)

vp_layer_for_inweb_no_text_sf(p, d_in, marker_col)

vp_layer_for_inweb_cbf_sf(p, d_in)

vp_layer_for_inweb_cbf_sf_gg(p, d_in, lab)

vp_layer_for_inweb_cbf_no_text_sf(p, d_in)
}
\description{
InWeb Layers for ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_inweb}()},
\code{\link{sp_layer_for_snp_to_gene_sgl}()},
\code{\link{sp_layer_for_uploaded_genes}()},
\code{\link{vp_layer_for_snp_to_gene_sgl}()},
\code{\link{vp_layer_for_uploaded_genes_gg}()}
}
\author{
April/Frederik
}
\concept{layers}
