% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_layer_for_snp_to_gene.R
\name{vp_layer_for_snp_to_gene_sgl}
\alias{vp_layer_for_snp_to_gene_sgl}
\alias{vp_layer_for_snp_to_gene_sgl_gg}
\alias{vp_layer_for_snp_to_gene_sgl_no_text}
\alias{vp_layer_for_snp_to_gene_sgl_cbf}
\alias{vp_layer_for_snp_to_gene_sgl_cbf_gg}
\alias{vp_layer_for_snp_to_gene_sgl_cbf_no_text}
\alias{vp_layer_for_snp_to_gene_mgl}
\alias{vp_layer_for_snp_to_gene_mgl_gg}
\alias{vp_layer_for_snp_to_gene_mgl_no_text}
\alias{vp_layer_for_snp_to_gene_mgl_cbf}
\alias{vp_layer_for_snp_to_gene_mgl_cbf_gg}
\alias{vp_layer_for_snp_to_gene_mgl_cbf_no_text}
\alias{vp_layer_for_snp_to_gene_none}
\alias{vp_layer_for_snp_to_gene_none_cbf}
\title{Layer for mapping SNPs to gene in ggplot2}
\usage{
vp_layer_for_snp_to_gene_sgl(p, snp_sgl, marker_col)

vp_layer_for_snp_to_gene_sgl_gg(p, snp_sgl, marker_col, lab)

vp_layer_for_snp_to_gene_sgl_no_text(p, snp_sgl, marker_col)

vp_layer_for_snp_to_gene_sgl_cbf(p, snp_sgl)

vp_layer_for_snp_to_gene_sgl_cbf_gg(p, snp_sgl, lab)

vp_layer_for_snp_to_gene_sgl_cbf_no_text(p, snp_sgl)

vp_layer_for_snp_to_gene_mgl(p, snp_mgl, marker_col)

vp_layer_for_snp_to_gene_mgl_gg(p, snp_mgl, marker_col, lab)

vp_layer_for_snp_to_gene_mgl_no_text(p, snp_mgl, marker_col)

vp_layer_for_snp_to_gene_mgl_cbf(p, snp_mgl)

vp_layer_for_snp_to_gene_mgl_cbf_gg(p, snp_mgl, lab)

vp_layer_for_snp_to_gene_mgl_cbf_no_text(p, snp_mgl)

vp_layer_for_snp_to_gene_none(p, d)

vp_layer_for_snp_to_gene_none_cbf(p, d)
}
\description{
Layer for mapping SNPs to gene in ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_inweb}()},
\code{\link{sp_layer_for_snp_to_gene_sgl}()},
\code{\link{sp_layer_for_uploaded_genes}()},
\code{\link{vp_layer_for_inweb}()},
\code{\link{vp_layer_for_uploaded_genes_gg}()}
}
\author{
April/Frederik
}
\concept{layers}
