% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_layer_for_uploaded_genes.R
\name{vp_layer_for_uploaded_genes_gg}
\alias{vp_layer_for_uploaded_genes_gg}
\title{Layer for uploaded genes in ggplot2}
\usage{
vp_layer_for_uploaded_genes_gg(p, d_g2s, usr_palette, lab)
}
\description{
Layer for uploaded genes in ggplot2
}
\seealso{
Other layers: 
\code{\link{sp_layer_for_inweb}()},
\code{\link{sp_layer_for_snp_to_gene_sgl}()},
\code{\link{sp_layer_for_uploaded_genes}()},
\code{\link{vp_layer_for_inweb}()},
\code{\link{vp_layer_for_snp_to_gene_sgl}()}
}
\author{
April/Frederik
}
\concept{layers}
